Restore previous state for libxml option
Whenever libxml_use_internal_errors() is called, the previous value for
it should be restored. This used to be the case in this piece of code,
but it was wrongly removed in e53bf5839b625e5ab92ffc4859331f4f46161bc6 ,
which has the nasty side effect of making the Validator component test
suite break with this message:

Validation failed: No DTD found!